+++ test/controllers/ControllerSpecBase.scala
@@ -25,15 +25,14 @@ import identifiers.register.establishers.individual.EstablisherDetailsId
 import identifiers.register.establishers.partnership.PartnershipDetailsId
 import identifiers.register.establishers.partnership.partner.PartnerDetailsId
 import identifiers.register.trustees.TrusteesId
-import identifiers.register.trustees.individual.{TrusteeDetailsId, TrusteeNewNinoId, TrusteeNinoId}
+import identifiers.register.trustees.individual.TrusteeDetailsId
 import identifiers.{AdviserNameId, SchemeNameId}
 import models.person.PersonDetails
-import models.{CompanyDetails, Nino, PartnershipDetails}
+import models.{CompanyDetails, PartnershipDetails}
 import org.joda.time.LocalDate
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import play.api.libs.json.Json
-import play.twirl.api.Html
 import utils.{Enumerable, MapFormats}
 
 trait ControllerSpecBase extends SpecBase with Enumerable.Implicits with MapFormats {
@@ -75,28 +74,6 @@ trait ControllerSpecBase extends SpecBase with Enumerable.Implicits with MapForm
         Json.obj(
           TrusteeDetailsId.toString ->
             PersonDetails("Test", Some("Trustee"), "Name", LocalDate.now)
-        )
-      )
-    )))
-
-  def getMandatoryTrusteeWithNewNinoId: FakeDataRetrievalAction = new FakeDataRetrievalAction(Some(
-    Json.obj(
-      "trustees" -> Json.arr(
-        Json.obj(
-          TrusteeDetailsId.toString ->
-            PersonDetails("Test", Some("Trustee"), "Name", LocalDate.now),
-          TrusteeNewNinoId.toString -> "CS121212C"
-        )
-      )
-    )))
-
-  def getMandatoryTrusteeWithOldNinoId: FakeDataRetrievalAction = new FakeDataRetrievalAction(Some(
-    Json.obj(
-      "trustees" -> Json.arr(
-        Json.obj(
-          TrusteeDetailsId.toString ->
-            PersonDetails("Test", Some("Trustee"), "Name", LocalDate.now),
-          TrusteeNinoId.toString -> Nino.Yes("CS121212C")
         )
       )
     )))
